# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ticketing Cinema API
  version: 1.0.0
  description: |
    Contract for the Ticketing Cinema system.

    The gateway is the single REST entry point. It forwards requests to
    internal services over TCP (user-service, cinema-service, ticket-service).
    This document only describes the public REST surface exposed by the gateway.

    ## Booking flow (two-step hold -> pay)
    1. `POST /bookings/holds` reserves seats with a Redis lock (5-minute TTL)
       and returns a `bookingId`, `paymentId` and `expiresAt`.
    2. `POST /bookings/{id}/pay` confirms the booking (mock payment). If the
       hold has expired the API returns `410 Gone` and the seats are released.
    3. `POST /bookings/{id}/cancel` releases the seats (only while PENDING).

    ## Roles
    - `admin` — full access to catalog management (movies, theaters, showtimes).
    - `user` — browse catalog, hold/pay/cancel bookings, view tickets.

    All prices are integer IDR (no decimals). All ids are UUID strings.
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development (gateway)
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid payload or email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      security: []
      tags: [Auth]
      summary: Login and receive an access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bookings/holds:
    post:
      tags: [Bookings]
      summary: Hold seats for a showtime (start of two-step booking)
      operationId: createBookingHold
      description: |
        Acquires a Redis lock on every requested seat. If any seat is already
        held or booked, no seats are locked and `409 Conflict` is returned.
        On success the booking is `PENDING` and expires at `expiresAt`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldRequest"
      responses:
        "201":
          description: Seats held, booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingHeld"
        "400":
          description: Invalid payload (empty seatIds, unknown seats, disabled seats)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Showtime not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: One or more seats are already held or booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatConflictError"
  /bookings:
    get:
      tags: [Bookings]
      summary: List the current user's bookings
      operationId: listMyBookings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          description: Filter by booking status
          schema:
            $ref: "#/components/schemas/BookingStatus"
      responses:
        "200":
          description: Paginated booking list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingPage"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Bookings]
      summary: Get a booking by id (owner or admin)
      operationId: getBooking
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the booking owner and not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}/pay:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Bookings]
      summary: Confirm a booking with a mock payment
      operationId: payBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayRequest"
      responses:
        "200":
          description: |
            Payment processed. Check `payment.status`:
            - `PAID` — booking is `CONFIRMED` and tickets are issued.
            - `FAILED` — booking remains `PENDING` (simulated payment failure),
              the client may retry until `expiresAt`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the booking owner and not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Booking is already CONFIRMED or CANCELLED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: Booking hold has expired; seats released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Bookings]
      summary: Cancel a pending booking and release its seats
      operationId: cancelBooking
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not the booking owner and not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Booking is not PENDING (cannot cancel CONFIRMED/EXPIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tickets/{code}:
    parameters:
      - name: code
        in: path
        required: true
        description: Ticket code (e.g. `TKT-A1B2C3`)
        schema:
          $ref: "#/components/schemas/TicketCode"
    get:
      security: []
      tags: [Tickets]
      summary: Look up an issued ticket by its code (public, for scanning)
      operationId: getTicketByCode
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token obtained from POST /auth/login.
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource id (MongoDB ObjectId)
      schema:
        $ref: "#/components/schemas/ObjectId"
    Page:
      name: page
      in: query
      description: Page number, 1-based
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      example: 66f3a1b2c3d4e5f6a7b8c9d0
    TicketCode:
      type: string
      pattern: "^TKT-[A-Z0-9]{6}$"
      example: TKT-A1B2C3
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          minimum: 0
          description: Price in IDR, no decimals
          example: 50000
        currency:
          type: string
          enum: [IDR]
    Error:
      type: object
      description: Standard error body (NestJS format)
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        error:
          type: string
          example: Not Found
        conflictSeatIds:
          type: array
          description: Present only on seat-conflict errors
          items:
            $ref: "#/components/schemas/ObjectId"
        paymentStatus:
          type: string
          description: Present only on payment errors
          enum: [PENDING, PAID, FAILED]
    SeatConflictError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          required: [conflictSeatIds]
          properties:
            conflictSeatIds:
              type: array
              items:
                $ref: "#/components/schemas/ObjectId"
            message:
              type: string
              example: One or more seats are already held or booked
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        email:
          type: string
          format: email
          example: budi@example.com
        name:
          type: string
          example: Budi
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: Stored hashed (bcrypt)
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          description: JWT (signed, contains userId and role)
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          description: Token lifetime in seconds
          example: 3600
        user:
          $ref: "#/components/schemas/User"
    AgeRating:
      type: string
      enum: [SU, BO, "13+", "17+", "21+"]
      description: "SU = all ages, BO = family/baby-friendly"
...
    SeatCategory:
      type: string
      enum: [REGULAR, VIP, COUPLE]
...
    MovieSummary:
      type: object
      required: [id, title, posterUrl, ageRating, durationMinutes]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        title:
          type: string
        posterUrl:
          type: [string, "null"]
          format: uri
        ageRating:
          $ref: "#/components/schemas/AgeRating"
        durationMinutes:
          type: integer
    TheaterSummary:
      type: object
      required: [id, name]
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
...
    BookingStatus:
      type: string
      enum: [PENDING, CONFIRMED, EXPIRED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [PENDING, PAID, FAILED]
    HoldRequest:
      type: object
      required: [showtimeId, seatIds]
      properties:
        showtimeId:
          $ref: "#/components/schemas/ObjectId"
        seatIds:
          type: array
          description: Seats to hold for this showtime
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/ObjectId"
          example: [66f3a1b2c3d4e5f6a7b8c9d1, 66f3a1b2c3d4e5f6a7b8c9d2]
    BookingSeat:
      type: object
      required: [seatId, row, number, category, price]
      properties:
        seatId:
          $ref: "#/components/schemas/ObjectId"
        row:
          type: string
          example: B
        number:
          type: integer
          example: 7
        category:
          $ref: "#/components/schemas/SeatCategory"
        price:
          $ref: "#/components/schemas/Money"
    Payment:
      type: object
      required: [paymentId, method, status, amount]
      properties:
        paymentId:
          type: string
          example: pay_01J2
        method:
          type: string
          enum: [MOCK]
        status:
          $ref: "#/components/schemas/PaymentStatus"
        amount:
          $ref: "#/components/schemas/Money"
        paidAt:
          type: [string, "null"]
          format: date-time
    Ticket:
      type: object
      required:
        - id
        - code
        - bookingId
        - showtimeId
        - movieTitle
        - theaterName
        - startsAt
        - seat
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        code:
          $ref: "#/components/schemas/TicketCode"
        bookingId:
          $ref: "#/components/schemas/ObjectId"
        showtimeId:
          $ref: "#/components/schemas/ObjectId"
        movieTitle:
          type: string
          example: The Grand Adventure
        theaterName:
          type: string
          example: Grand Cineplex 1
        startsAt:
          type: string
          format: date-time
        seat:
          type: object
          required: [seatId, row, number]
          properties:
            seatId:
              $ref: "#/components/schemas/ObjectId"
            row:
              type: string
            number:
              type: integer
    Booking:
      type: object
      required:
        - id
        - userId
        - showtimeId
        - movie
        - theater
        - startsAt
        - seats
        - totalAmount
        - status
        - expiresAt
        - payment
        - createdAt
        - updatedAt
      properties:
        id:
          $ref: "#/components/schemas/ObjectId"
        userId:
          $ref: "#/components/schemas/ObjectId"
        showtimeId:
          $ref: "#/components/schemas/ObjectId"
        movie:
          $ref: "#/components/schemas/MovieSummary"
        theater:
          $ref: "#/components/schemas/TheaterSummary"
        startsAt:
          type: string
          format: date-time
        seats:
          type: array
          items:
            $ref: "#/components/schemas/BookingSeat"
        totalAmount:
          $ref: "#/components/schemas/Money"
        status:
          $ref: "#/components/schemas/BookingStatus"
        expiresAt:
          type: string
          format: date-time
          description: Hold deadline; past this the booking expires and locks release
        payment:
          $ref: "#/components/schemas/Payment"
        tickets:
          type: array
          description: Present when status is CONFIRMED
          items:
            $ref: "#/components/schemas/Ticket"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookingHeld:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          required: [bookingId, seats]
          properties:
            bookingId:
              $ref: "#/components/schemas/ObjectId"
              description: Same as `id`; included for clarity at creation time
    BookingPage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    PayRequest:
      type: object
      required: [paymentMethod]
      properties:
        paymentMethod:
          type: string
          enum: [MOCK]
        simulate:
          type: string
          description: Mock provider outcome (defaults to SUCCESS)
          enum: [SUCCESS, FAILURE]
          default: SUCCESS
    PaginationMeta:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3